refactor(nosskey-sdk): share base64 helpers in utils
bytesToBase64 / base64ToBytes were duplicated between nip44.ts and
nip04.ts with the same browser-vs-Node fallback. Hoist the generic
codec into utils.ts and keep the NIP-44-specific version-prefix and
empty-string checks in a thin decodePayload() wrapper.
