+++ packages/nosskey-sdk/src/utils.ts
@@ -20,6 +20,33 @@ export function bytesToHex(bytes: Uint8Array): string {
 }
 
 /**
+ * バイト配列を base64 文字列に変換する。ブラウザの `btoa` を優先し、
+ * Node 環境では `Buffer` にフォールバックする。
+ */
+export function bytesToBase64(bytes: Uint8Array): string {
+  if (typeof btoa === 'function') {
+    let bin = '';
+    for (let i = 0; i < bytes.length; i++) bin += String.fromCharCode(bytes[i]);
+    return btoa(bin);
+  }
+  return Buffer.from(bytes).toString('base64');
+}
+
+/**
+ * base64 文字列をバイト配列に変換する。`atob` がない環境では `Buffer`
+ * にフォールバックする。
+ */
+export function base64ToBytes(str: string): Uint8Array {
+  if (typeof atob === 'function') {
+    const bin = atob(str);
+    const bytes = new Uint8Array(bin.length);
+    for (let i = 0; i < bin.length; i++) bytes[i] = bin.charCodeAt(i);
+    return bytes;
+  }
+  return new Uint8Array(Buffer.from(str, 'base64'));
+}
+
+/**
  * 16進数文字列をバイト配列に変換する
  * @param hex 変換する16進数文字列
  * @returns バイト配列
